switch all queries to use react-query
For now no UI elements have been chaged. For suggested future work, it would make sense to pass the `QueryStatus` objects related
to a stat directly to the `StatsBox`. The `StatsBox` can then be modified to show live reload, errors, or force reload.

`useContext` for SNX and sUSD were replaced with `useSNXInfo` and `useSUSDInfo`, which performs the necessary queries to populate the data.
since `react-query` maintains its own internal cache, there is no need to use context.

The default auto-refresh period used by `react-query` might be a bit to aggressive for the purposes of this dashboard, particularly if an API goes
down or starts producing exceptions.
Please advise on if this should be adjusted.